[US1302] More code to support download q cookie.
Currently there's a bug that I'm still have to fix. If you leave the page and then come back the q will still be there
except you can't remove anything from it. This bug is a result of the way the download q is coded to remove items 
from itself. It looks up a particular id from the search results datatable to do some stuff with it except that when you 
get back to the page the search results list does not exist, so it simply can't remove it from the q. 

Still needs a little bit of work