Updated Unit tests and Implement in vertex
There seems to be some slowness in loading generating the datatable. This probably has to do with using jQueryUI and how
the table is being generated and passed to the datatable.

Once generated and rendered though it is much faster at filtering. There are some css styling issues that need to be worked out. Also some of the event bindings are broken.
The map tiles are not highlighted when a hover event occurs